flask_app/requirements.txt
alembic==1.6.5
bcrypt==3.2.0
blinker==1.4
cffi==1.14.6
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gcloud==0.17.0
googleapis-common-protos==1.53.0
greenlet==1.1.1
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
jws==0.1.3
Mako==1.1.5
MarkupSafe==2.0.1
names==0.3.0
oauth2client==3.0.0
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug==2.0.1
WTForms==2.3.3
